Refactored more code in QuizForm.cs
